Fixes for better integration with Jupyter in base environment.
We need to set JUPYTER_PATH in addition to other variables so JupyterLab
extensions are correctly found after installation.

Also add proper configuration of the pip installation location inside the conda
base environment, so the overlay works as such also for the base env. This
means that a full removal of conda will also nuke the overlay (in conda/copip).